+++ shared/lib.rs
@@ -1,3 +1,4 @@
+pub mod log;
 pub mod message;
 pub mod protocol;
 pub mod settings;